refactoring and new tests
