+++ src/main/java/UI/UserRegistrationResponseFormatter.java
@@ -1,4 +1,7 @@
-package use_cases.user_registration;
+package UI;
+
+import use_cases.user_registration.UserRegistrationPresenter;
+import use_cases.user_registration.UserRegistrationResponse;
 
 public class UserRegistrationResponseFormatter implements UserRegistrationPresenter {
